Refactor main.rs into smaller modules
- Split i18n functionality into i18n module
- Split CLI options into cli/options.rs
- Split help display into cli/help.rs
- Split interactive console into cli/console.rs
- Split server management into cli/server.rs
- Update main.rs to use new modules
- Improve code organization and maintainability
