+++ src/cli/console.rs
@@ -1,0 +1,116 @@
+use anyhow::Result;
+use console::{style, Term};
+use dialoguer::{theme::ColorfulTheme, Select};
+use crate::i18n::LocaleConfig;
+
+#[derive(Debug, Clone)]
+pub struct AppState {
+    pub api_port: u16,
+    pub frontend_port: u16,
+    pub locale: LocaleConfig,
+    pub api_url: String,
+    pub frontend_url: String,
+}
+
+pub struct InteractiveConsole;
+
+impl InteractiveConsole {
+    pub fn display_logo() {
+        println!("{}", style(r#"
+    ╭──────────────────────────────────────╮
+    │                                      │
+    │   ╭─────────╮  RUSTORIUM            │
+    │   │ ₿ │ ⟠ │ │                       │
+    │   │ ◎ │ ₳ │ │  Blockchain Platform  │
+    │   ╰─────────╯                       │
+    │                                      │
+    ╰──────────────────────────────────────╯
+    "#).cyan());
+    }
+
+    pub async fn run(app_state: &AppState) -> Result<()> {
+        let term = Term::stdout();
+        term.clear_screen()?;
+        
+        Self::display_logo();
+
+        // 言語選択
+        let languages = vec!["English", "日本語", "中文", "한국어"];
+        let language_selection = Select::with_theme(&ColorfulTheme::default())
+            .with_prompt(style("Select your language / 言語を選択 / 选择语言 / 언어 선택").cyan().bold().to_string())
+            .items(&languages)
+            .default(0)
+            .interact()?;
+
+        let language_code = match language_selection {
+            0 => "en",
+            1 => "ja",
+            2 => "zh",
+            3 => "ko",
+            _ => "en",
+        };
+
+        let locale = LocaleConfig::new(language_code);
+        println!("\n{}", style(locale.get_message("welcome")).bold());
+        println!("\n{}", style("🌐 Services:").cyan().bold());
+        println!("  API: {}", style(&app_state.api_url).underlined());
+        println!("  Frontend: {}", style(&app_state.frontend_url).underlined());
+        println!();
+
+        loop {
+            let actions = vec![
+                locale.get_message("account"),
+                locale.get_message("transaction"),
+                locale.get_message("smart_contract"),
+                locale.get_message("blockchain"),
+                locale.get_message("settings"),
+                locale.get_message("exit"),
+            ];
+
+            let selection = Select::with_theme(&ColorfulTheme::default())
+                .with_prompt(style(locale.get_message("select_action")).cyan().bold().to_string())
+                .items(&actions)
+                .default(0)
+                .interact()?;
+
+            match selection {
+                0 => Self::handle_account_management(&locale).await?,
+                1 => Self::handle_transactions(&locale).await?,
+                2 => Self::handle_smart_contracts(&locale).await?,
+                3 => Self::handle_blockchain_info(&locale).await?,
+                4 => Self::handle_settings(&locale).await?,
+                5 => break,
+                _ => {}
+            }
+
+            println!();
+        }
+
+        Ok(())
+    }
+
+    async fn handle_account_management(_locale: &LocaleConfig) -> Result<()> {
+        // アカウント管理の実装
+        Ok(())
+    }
+
+    async fn handle_transactions(_locale: &LocaleConfig) -> Result<()> {
+        // トランザクション管理の実装
+        Ok(())
+    }
+
+    async fn handle_smart_contracts(_locale: &LocaleConfig) -> Result<()> {
+        // スマートコントラクト管理の実装
+        Ok(())
+    }
+
+    async fn handle_blockchain_info(_locale: &LocaleConfig) -> Result<()> {
+        // ブロックチェーン情報の実装
+        Ok(())
+    }
+
+    async fn handle_settings(_locale: &LocaleConfig) -> Result<()> {
+        // 設定の実装
+        Ok(())
+    }
+}
