+++ src/cli/mod.rs
@@ -1,0 +1,9 @@
+mod console;
+mod help;
+mod options;
+mod server;
+
+pub use console::{AppState, InteractiveConsole};
+pub use help::print_help;
+pub use options::AppOptions;
+pub use server::ServerManager;
